+++ tests/unit/Models.Bom.spec.js
@@ -87,10 +87,12 @@ suite('Models.Bom', () => {
       test(`for: ${newVersion}`, () => {
         const bom = new Bom()
         assert.notStrictEqual(bom.version, newVersion)
-
-        assert.throws(() => {
-          bom.version = newVersion
-        }, /not PositiveInteger/i)
+        assert.throws(
+          () => {
+            bom.version = newVersion
+          },
+          /not PositiveInteger/i
+        )
       })
     )
   )