+++ HISTORY.md
@@ -24,11 +24,14 @@ All notable changes to this project will be documented in this file.
     * `Vulnerability`, `VulnerabilityRepository`
 * Docs
   * Use [TSDoc](https://tsdoc.org/) syntax in TypeScript files, instead of [JSDoc](https://jsdoc.app/) (via [#318], [#453])
+* Misc
+  * Added tests for internal helpers (via [#454])
 
 [#164]: https://github.com/CycloneDX/cyclonedx-javascript-library/issues/164
 [#318]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/318
 [#419]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/419
 [#453]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/453
+[#454]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/454
 
 ## 1.10.0 - 2023-01-28